Rename trim-whitespace-core
