+++ mixins/trim-whitespace/trim-whitespace-core.js
